+++ imagetest/cmd/manager/main.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest"
 	imagevalidation "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/image_validation"
-	imageboot "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/imageboot"
+	imageboot "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/image_boot"
 )
 
 var (
