+++ imagetest/test_suites/image_boot/setup.go
@@ -1,0 +1,16 @@
+package imageboot
+
+import "github.com/GoogleCloudPlatform/guest-test-infra/imagetest"
+
+// Name is the name of the test package. It must match the directory name.
+var Name = "image-boot"
+
+// TestSetup sets up the test workflow.
+func TestSetup(t *imagetest.TestWorkflow) error {
+	rebootVM, err := t.CreateTestVM("reboot-test")
+	if err != nil {
+		return err
+	}
+	rebootVM.Reboot()
+	return nil
+}
